mux: change mechanism for flushing on protocol termination
The mux ensures that when a protocol thread terminates normally (rather
than via an exception) that it waits for outbound messages to be sent.
This allows for a clean shutdown for the remote peer and this is
required in various tests.

The previous mechanism used a counter that was shared across all the
mini-protocols in the mux bundle, and had to be incremented and
decremented in vairious places.

The new mechanism is much simpler because it is not shared across all
the mini-protocols. It is per-protocol. It simply waits for the outbound
queue for that protocol to become empty. It should also be slightly
faster since there are fewer STM transactions and allocations in the
outbound data path.